+++ R/remove_labels.R
@@ -139,11 +139,11 @@ remove_user_na.haven_labelled_spss <- function(x,
                                                user_na_to_na = FALSE,
                                                user_na_to_tagged_na = FALSE) {
   if (user_na_to_tagged_na) {
-    if (typeof(x) == "character") {
-      stop(
-        "'user_na_to_tagged_na' cannot be used with character labelled vectors."
-      )
-    }
+    if (typeof(x) == "character")
+      cli::cli_abort(paste(
+        "{.fn user_na_to_tagged_na} cannot be used with",
+        "character labelled vectors."
+      ))
 
     val_to_tag <- x[is.na(x) & !is.na(unclass(x))] %>%
       unclass() %>%
@@ -151,18 +151,20 @@ remove_user_na.haven_labelled_spss <- function(x,
       sort()
 
     if (length(val_to_tag) > 26) {
-      warning(
-        length(val_to_tag),
-        " different user-defined missing values found in 'x'.\n",
-        "A maximum of 26 could be tagged.\n",
-        "'user_na_to_tagged_na' has been ignored.\n",
-        "'user_na_to_na = TRUE' has been used instead."
-      )
+      cli::cli_warn(c(
+        "{length(val_to_tag)} different user-defined missing
+         values found in 'x'.",
+        i = "A maximum of 26 could be tagged.",
+        "!" = "'user_na_to_tagged_na' has been ignored.",
+        i = "'user_na_to_na = TRUE' has been used instead."
+      ))
       user_na_to_na <- TRUE
     } else {
       if (is.integer(x)) {
         x <- as.double(unclass(x)) %>% copy_labels_from(x)
-        message("'x' has been converted into a double vector.")
+        cli::cli_inform(c(
+          i = "{.arg x} has been converted into a double vector."
+        ))
         val_to_tag <- as.double(val_to_tag)
       }