+++ R/to_labelled.R
@@ -97,11 +97,7 @@ foreign_to_labelled <- function(x) {
   # if imported with read.spss(to.data.frame=FALSE) it's a
   # list, not a df
   if (!is.data.frame(x)) {
-    if (requireNamespace("dplyr")) {
-      x <- dplyr::as_tibble(x)
-    } else {
-      x <- as.data.frame(x, stringsAsFactors = FALSE)
-    }
+    x <- dplyr::as_tibble(x)
   }
 
   # variable labels (read.spss)
@@ -173,11 +169,7 @@ memisc_to_labelled <- function(x) {
     return(x)
   }
 
-  if (!requireNamespace("memisc")) {
-    stop("memisc package is required to convert a data.set",
-      call. = FALSE, domain = "R-labelled"
-    )
-  }
+  rlang::check_installed("memisc", "to convert a data.set.")
 
   df <- as.data.frame(x)
   for (var in names(x)) {
@@ -257,7 +249,7 @@ to_labelled.factor <- function(x, labels = NULL, .quiet = FALSE, ...) {
           all(!is.na(l$code)) &&
           all(!is.na(l$code))
       ) {
-        warning("'x' looks prefixed, but duplicated codes found.")
+        cli::cli_warn("{.arg x} looks prefixed, but duplicated codes found.")
       }
       # normal case
       labs <- seq_along(levels(x))
@@ -267,7 +259,7 @@ to_labelled.factor <- function(x, labels = NULL, .quiet = FALSE, ...) {
       # "[code] label" case
       num_l <- suppressWarnings(as.numeric(l$code))
       if (!.quiet && all(!is.na(num_l)) && any(duplicated(num_l))) {
-        warning("All codes seem numeric but some duplicates found.")
+        cli::cli_warn("All codes seem numeric but some duplicates found.")
       }
       if (all(!is.na(num_l)) && !any(duplicated(num_l))) {
         l$code <- as.numeric(l$code)