+++ R/var_label.R
@@ -152,15 +152,12 @@ var_label.data.frame <- function(x,
   if (
     (!is.character(value) && !is.null(value)) && !is.list(value) ||
       (is.character(value) && length(value) > 1 && length(value) != ncol(x))
-  ) {
-    stop(
-      paste0(
-        "`value` should be a named list, NULL, a single character string or a ",
-        "character vector of same length than the number of columns in `x`"
-      ),
-      call. = FALSE, domain = "R-labelled"
-    )
-  }
+  )
+    cli::cli_abort(paste(
+      "{.arg value} should be a named list, NULL, a single character string",
+      "or a character vector of same length than the number of columns",
+      "in {.arg x}."
+    ))
   if (is.character(value) && length(value) == 1) {
     value <- as.list(rep(value, ncol(x)))
     names(value) <- names(x)
@@ -178,11 +175,11 @@ var_label.data.frame <- function(x,
   }
 
   if (!all(names(value) %in% names(x))) {
-    missing_names <- stringr::str_c(
-      setdiff(names(value), names(x)),
-      collapse = ", "
-    )
-    stop("some variables not found in x:", missing_names)
+    missing_names <- setdiff(names(value), names(x))
+
+    cli::cli_abort(c(
+      "Can't find variables {.var {missing_names}}  in {.arg x}."
+    ))
   }
 
   value <- value[names(value) %in% names(x)]
@@ -263,17 +260,20 @@ set_variable_labels <- function(.data, ..., .labels = NA, .strict = TRUE) {
   if (!identical(.labels, NA)) {
     if (!.strict) {
       .labels <- .labels[intersect(names(.labels), names(.data))]
+    } else {
+      check_character(.labels)
+
     }
     var_label(.data) <- .labels
   }
   values <- rlang::dots_list(...)
   if (length(values) > 0) {
     if (.strict && !all(names(values) %in% names(.data))) {
-      missing_names <- stringr::str_c(
-        setdiff(names(values), names(.data)),
-        collapse = ", "
-      )
-      stop("some variables not found in .data: ", missing_names)
+      missing_names <- setdiff(names(values), names(.data))
+
+      cli::cli_abort(c(
+        "Can't find variables {.var {missing_names}}  in {.arg .data}."
+      ))
     }
 
     for (v in intersect(names(values), names(.data))) {
@@ -299,13 +299,7 @@ get_label_attribute <- function(x) {
 #' @rdname var_label
 #' @export
 set_label_attribute <- function(x, value) {
-  if ((!is.character(value) && !is.null(value)) || length(value) > 1) {
-    stop(
-      "`value` should be a single character string or NULL",
-      call. = FALSE,
-      domain = "R-labelled"
-    )
-  }
+  check_string(value, allow_null = TRUE)
   attr(x, "label") <- value
   x
 }