+++ NEWS.md
@@ -1,3 +1,5 @@
+# labelled (development version)
+
 # labelled 2.13.0
 
 **New features**