+++ NEWS.md
@@ -6,6 +6,10 @@
   name inside `.fn` by using `names()` (#163)
 * `var_label()` gets new options `"na"` and `"empty"` for `null_action`
 
+**Improvements**
+
+* systematic use of `{cli}` for errors, warnings and messages (#167)
+
 # labelled 2.13.0
 
 **New features**